backend-api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
pyjwt
python-multipart
python-dotenv